# back/openapi.yaml
openapi: 3.0.3
info:
  title: IS Course Work Back API
  description: >-
    REST API that powers the restaurant project. It exposes
    endpoints to manage employees, dishes, ingredients, tables, orders, bills,
    feedback, reports, and employee wallets. Most operations require a JWT that
    encodes the employee id and position (role).
  version: 1.0.0
servers:
  - url: http://localhost:8080/is_coursework/api
    description: Local application server
...
paths:
...
  /employee/modify:
    post:
      tags: [Employees]
      summary: Update employee data
      description: Only administrators may modify employees.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '200':
          description: Employee updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /employee/delete/{id}:
    delete:
      tags: [Employees]
      summary: Remove an employee
      description: Only administrators may delete employees.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Employee deleted
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /employee/get-all:
    get:
      tags: [Employees]
      summary: Get all employees
      description: Only administrators may view the directory.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /feedback/create:
    post:
      tags: [Feedback]
      summary: Leave feedback for a table
      description: Guests can submit feedback without authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreationRequest'
      responses:
        '201':
          description: Feedback stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackCreationRequest'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
  /feedback/get-all:
    get:
      tags: [Feedback]
      summary: View all feedbacks
      description: Only administrators may read all feedback.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Feedback entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /feedback/get-for-employee/{employeeId}:
    get:
      tags: [Feedback]
      summary: View feedback that for employee
      description: Available to administrators and to the employee themselves.
      security:
        - BearerAuth: []
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Feedback entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/get-for-dish/{id}:
    get:
      tags: [Ingredients]
      summary: List ingredients for a dish
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/create:
    post:
      tags: [Ingredients]
      summary: Create an ingredient
      description: Only administrators may create ingredients.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ingredient'
      responses:
        '200':
          description: Ingredient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/modify:
    post:
      tags: [Ingredients]
      summary: Update an ingredient
      description: Only administrators may edit ingredients.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Ingredient'
      responses:
        '200':
          description: Ingredient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/delete/{id}:
    delete:
      tags: [Ingredients]
      summary: Delete an ingredient
      description: Only administrators may delete ingredients.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Ingredient deleted
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/get-all:
    get:
      tags: [Ingredients]
      summary: Get all ingredients
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Ingredients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/increase-amount:
    post:
      tags: [Ingredients]
      summary: Increase ingredient amount
      description: Only administrators may increase ingredient amount.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientAmountUpdate'
      responses:
        '200':
          description: Amount updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /ingredient/reset-cost:
    post:
      tags: [Ingredients]
      summary: Reset ingredient cost
      description: Only administrators may adjust cost.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientCostUpdate'
      responses:
        '200':
          description: Cost updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: The caller does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
      additionalProperties: false
...
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        positions:
          $ref: '#/components/schemas/Positions'
        name:
          type: string
        secondName:
          type: string
        patronymic:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
        sex:
          type: string
          maxLength: 1
      required: [positions, name, secondName, birthDate, sex]
    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        positions:
          $ref: '#/components/schemas/Positions'
        name:
          type: string
        secondName:
          type: string
        patronymic:
          type: string
          nullable: true
      required: [id, positions, name, secondName]
    FeedbackCreationRequest:
      type: object
      properties:
        tableNumber:
          $ref: '#/components/schemas/TableNumber'
        rating:
          type: integer
          format: int32
          minimum: 1
          maximum: 5
        comment:
          type: string
        tipAmount:
          type: number
          format: double
      required: [tableNumber, rating]
    FeedbackResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        time:
          type: string
          format: date-time
        journalLog:
          $ref: '#/components/schemas/JournalLogResponse'
        rating:
          type: integer
          format: int32
        comment:
          type: string
        tipAmount:
          type: number
          format: double
      required: [id, time, journalLog, rating]
    Ingredient:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        price:
          type: number
          format: double
          description: Price per 100 grams
        amount:
          type: number
          format: double
          description: Amount in stock (grams)
      required: [name, price, amount]
    IngredientAmountUpdate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        amount:
          type: number
          format: double
          description: Amount to set for the ingredient
      required: [id, amount]
    IngredientCostUpdate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cost:
          type: number
          format: double
      required: [id, cost]
...
    JournalLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tableStatus:
          $ref: '#/components/schemas/TableStatus'
        employee:
          $ref: '#/components/schemas/EmployeeResponse'
        tableNumber:
          $ref: '#/components/schemas/TableNumber'
        time:
          type: string
          format: date-time
      required: [id, tableStatus, employee, tableNumber, time]
...
    TableNumber:
      type: string
      enum: [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10]
    TableStatus:
      type: string
      enum: [free, occupied, not_paid, paid]
...
    Positions:
      type: string
      enum: [waiter, cook, barman, admin]